Don't use unsigned short for pixel coordinates, it causes a headache and SDL does clipping for us anyway

+++ gc.h
@@ -22,7 +22,7 @@ public:
 	}
 
 	virtual void set_color(unsigned char r, unsigned char g, unsigned char b) = 0;
-	virtual void box(unsigned short x, unsigned short y, unsigned short w, unsigned short h) = 0;
+	virtual void box(int x, int y, int w, int h) = 0;
 
 	virtual void prepare() = 0;
 	virtual void commit() = 0;
