+++ sdlgc.cpp
@@ -15,7 +15,7 @@ void sdlgc::set_color(unsigned char r, unsigned char g, unsigned char b)
 	color = SDL_MapRGB(srf->format, r, g, b);
 }
 
-void sdlgc::box(unsigned short x, unsigned short y, unsigned short w, unsigned short h)
+void sdlgc::box(int x, int y, int w, int h)
 {
 	SDL_Rect r;
